requirements.txt
# LinkedIn JD Analyzer - Dependencies

# Web scraping
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.31.0

# Data processing
pandas==2.2.0

# Email (Gmail API with OAuth 2.0)
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Scheduling
schedule==1.2.1
apscheduler==3.10.4

# AI/NLP for analysis
openai==1.12.0

# Environment variables
python-dotenv==1.0.1

# HTML template
jinja2==3.1.3
